requirements.txt
streamlit>=1.28.0
snowflake-connector-python>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
scikit-learn>=1.3.0
plotly>=5.18.0
python-dotenv>=1.0.0
newsapi-python>=0.2.7
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.0
langchain>=0.1.0
tqdm>=4.65.0
trulens-eval>=0.18.0
